feat: change beatmap search options ui
