+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom"
-import { Beatmaps, Download, Info } from './Pages'
+import { Beatmaps, Download, Info, NotFound } from './Pages'
 
 function App() {
     return(
@@ -11,6 +11,8 @@ function App() {
                 <Route path="/dev" element={<Beatmaps dev={true}/>}/>
                 <Route path="/d/:beatmapsetid" element={<Download />}/>
                 <Route path="/info" element={<Info />}/>
+
+                <Route path="*" element={<NotFound />}/>
             </Routes>
         </BrowserRouter>
     )
