+++ src/Components/Searchbar.js
@@ -14,25 +14,8 @@ function Searchbar() {
     const { Search } = Input
 
     const [apiJson] = useGlobalState("apiJson")
-    const [detailSearchTmp] = useGlobalState("detailSearchTmp")
-    const [detailSearch] = useGlobalState("detailSearch")
 
     const [tmp, setTmp] = useState(0)
-    const [detailSearchChange, setDetailSearchChange] = useState(false)
-
-    const [globalDirectDownload] = useGlobalState("downloadDirect")
-    const [globalNoVideo] = useGlobalState("globalNoVideo")
-    const [globalNoBg] = useGlobalState("globalNoBg")
-    const [globalNoHitsound] = useGlobalState("globalNoHitsound")
-    const [globalNoStoryboard] = useGlobalState("globalNoStoryboard")
-
-    const [zipList] = useGlobalState("zipList")
-
-    const { x } = useSpring({
-        from: { x: 0 },
-        x: detailSearch ? 1 : 0,
-        config: { duration: 100 },
-      })
 
     useEffect(() => {
         // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -53,43 +36,6 @@ function Searchbar() {
         
         setTmp(new Date().getMilliseconds())
         // searchParamHandler("q", apiJson.query)
-    }
-
-    function searchbarTypeHandler(event, target) {
-        event.stopPropagation()
-        event.preventDefault()
-
-        if (target !== true) {
-            apiJson.ar.min = 0
-            apiJson.ar.max = 0
-            apiJson.cs.min = 0
-            apiJson.cs.max = 0
-            apiJson.accuracy.min = 0
-            apiJson.accuracy.max = 0
-            apiJson.drain.min = 0
-            apiJson.drain.max = 0
-            apiJson.bpm.min = 0
-            apiJson.bpm.max = 0
-            apiJson.difficultyRating.min = 0
-            apiJson.difficultyRating.max = 0
-        }
-        if (target === true) {
-            apiJson.ar.min = detailSearchTmp.ar.min
-            apiJson.ar.max = detailSearchTmp.ar.max
-            apiJson.cs.min = detailSearchTmp.cs.min
-            apiJson.cs.max = detailSearchTmp.cs.max
-            apiJson.accuracy.min = detailSearchTmp.accuracy.min
-            apiJson.accuracy.max = detailSearchTmp.accuracy.max
-            apiJson.drain.min = detailSearchTmp.drain.min
-            apiJson.drain.max = detailSearchTmp.drain.max
-            apiJson.bpm.min = detailSearchTmp.bpm.min
-            apiJson.bpm.max = detailSearchTmp.bpm.max
-            apiJson.difficultyRating.min = detailSearchTmp.difficultyRating.min
-            apiJson.difficultyRating.max = detailSearchTmp.difficultyRating.max
-        }
-
-        setGlobalState("detailSearch", target)
-        setTimeout(() => { requestNewBeatmapData(false) }, 250);
     }
 
     function searchParamHandler(target, value) {
@@ -123,220 +69,6 @@ function Searchbar() {
         window.history.replaceState("", document.title, uri);
     }
 
-    function searchbarOptionChangeHandler(event, target, value) {
-        event.stopPropagation()
-        event.preventDefault()
-        if (target === "ranked") {
-            if (value === "any" || value === "") {
-                apiJson[target] = value
-                if (value === "any") apiJson[target] = "all"
-            }
-            else {
-                apiJson[target] = apiJson[target].replace('all', '')
-                if (apiJson[target].includes(value)) {
-                    if (apiJson[target][apiJson[target].indexOf(value) - 1] === ',')
-                        apiJson[target] = apiJson[target].replace(`,${value}`, "")
-                    else apiJson[target] = apiJson[target].replace(value, "")
-                } else {
-                    if (apiJson[target] === "") apiJson[target] += value
-                    else apiJson[target] += `,${value}`
-                }
-                do {
-                    if (apiJson[target][0] === ",") apiJson[target] = apiJson[target].slice(1)
-                } while (apiJson[target][0] === ",")
-            }
-        } else {
-            if (apiJson[target] === value) return
-            apiJson[target] = value
-        }
-
-        setTmp(new Date().getMilliseconds())
-        searchParamHandler(target === "ranked" ? "s" : target, apiJson[target] === "all" ? "any" : apiJson[target])
-        requestNewBeatmapData(false)
-    }
-
-    function searchbarOptionChangeHandlerForSortBy(event, target) {
-        event.stopPropagation()
-        event.preventDefault()
-        if (apiJson.sort.includes(target)) {
-            if (apiJson.sort.includes('desc')) apiJson.sort = `${target}_asc`
-            else apiJson.sort = `${target}_desc`
-        } else apiJson.sort = `${target}_desc`
-
-        setTmp(new Date().getMilliseconds())
-        searchParamHandler('sort', apiJson.sort === "ranked_desc" ? "": apiJson.sort )
-        requestNewBeatmapData(false)
-    }
-
-    function searchbarOptionChangeHandlerForExtra(event, target) {
-        event.stopPropagation()
-        event.preventDefault()
-        if (target === "video") {
-            if (apiJson.extra.includes('video')) {
-                if (apiJson.extra.includes('storyboard')) apiJson.extra = apiJson.extra.replace('.video', '')
-                else apiJson.extra = apiJson['extra'].replace('video', '')
-            } else if ((apiJson.extra.includes('storyboard'))) apiJson.extra += ".video"
-            else apiJson.extra += "video"
-        } else 
-        if (apiJson.extra.includes('storyboard')) {
-            if (apiJson.extra.includes('video')) apiJson.extra = apiJson.extra.replace('storyboard.', '')
-            else apiJson.extra = apiJson.extra.replace('storyboard', '')
-        } else if (apiJson.extra.includes('video')) apiJson.extra = 'storyboard.video'
-        else apiJson.extra = 'storyboard'
-
-        setTmp(new Date().getMilliseconds())
-        searchParamHandler('e', apiJson.extra)
-        requestNewBeatmapData(false)    
-    }
-
-    function searchbarOptionChangeHandlerForDownload(event) {
-        event.stopPropagation()
-        event.preventDefault()
-
-        setGlobalState("downloadDirect", !globalDirectDownload)
-
-        setTmp(new Date().getMilliseconds())
-    }
-
-    function searchbarOptionChangeHandlerForDlOptions(event, target) {
-        event.stopPropagation()
-        event.preventDefault()
-        switch (target) {
-            case "video":
-                setGlobalState("globalNoVideo", !globalNoVideo)
-                break
-            case "background":
-                setGlobalState("globalNoBg", !globalNoBg)
-                break
-            case "hitsound":
-                setGlobalState("globalNoHitsound", !globalNoHitsound)
-                break
-            case "storyboard":
-                setGlobalState("globalNoStoryboard", !globalNoStoryboard)
-                break
-            default:
-                break
-        }
-
-        setTmp(new Date().getMilliseconds())
-    }
-
-    function zipDownloadHandler(e, selectedOnly=true) {
-        e.stopPropagation()
-        e.preventDefault()
-        if (zipList.length === 0 && selectedOnly) {
-            return message.warning("Not Seleceted to download")
-        }
-        message.info("Please wait to download ...", 10)
-        if (!selectedOnly) {
-            var paramList = []
-            if (globalNoVideo) paramList.push(true)
-            else paramList.push(false)
-            if (globalNoBg) paramList.push(true)
-            else paramList.push(false)
-            if (globalNoHitsound) paramList.push(true)
-            else paramList.push(false)
-            if (globalNoStoryboard) paramList.push(true)
-            else paramList.push(false)
-            var tmp = []
-            getGlobalState("apiResult").map(function(bmap, i) {
-                if (i < 30) {
-                    var url = GeneralMixins.generateDownloadURL(bmap.id, paramList)
-                    var bname = `${bmap.id} ${bmap.artist} - ${bmap.title}.osz`
-                    tmp.push({'name': bname, 'url': url})
-                }
-            })
-            setGlobalState("zipList", tmp)
-        }
-        GeneralMixins.zipDownloadHandler()
-    }
-
-    const detailSearchbarValueHandlerForAR = (value) => {
-        apiJson.ar.min = value[0]
-        apiJson.ar.max = value[1]
-
-        detailSearchTmp.ar.min = value[0]
-        detailSearchTmp.ar.max = value[1]
-    }
-
-    const detailSearchbarValueHandlerForCS = (value) => {
-        apiJson.cs.min = value[0]
-        apiJson.cs.max = value[1]
-
-        detailSearchTmp.cs.min = value[0]
-        detailSearchTmp.cs.max = value[1]
-    }
-
-    const detailSearchbarValueHandlerForOD = (value) => {
-        apiJson.accuracy.min = value[0]
-        apiJson.accuracy.max = value[1]
-
-        detailSearchTmp.accuracy.min = value[0]
-        detailSearchTmp.accuracy.max = value[1]
-    }
-
-    const detailSearchbarValueHandlerForHP = (value) => {
-        apiJson.drain.min = value[0]
-        apiJson.drain.max = value[1]
-
-        detailSearchTmp.drain.min = value[0]
-        detailSearchTmp.drain.max = value[1]
-    }
-
-    const detailSearchbarValueHandlerForBPM = (value) => {
-        apiJson.bpm.min = value[0]
-        apiJson.bpm.max = value[1]
-
-        detailSearchTmp.bpm.min = value[0]
-        detailSearchTmp.bpm.max = value[1]
-    }
-
-    const detailSearchbarValueHandlerForSR = (value) => {
-        apiJson.difficultyRating.min = value[0]
-        apiJson.difficultyRating.max = value[1]
-
-        detailSearchTmp.difficultyRating.min = value[0]
-        detailSearchTmp.difficultyRating.max = value[1]
-    }
-
-    const detailSearchbarValueAfterChangeHandler = (value) => {
-        setDetailSearchChange(true)
-        clearTimeout(delay)
-        delay = setTimeout(function() {
-            requestNewBeatmapData(false)
-        }, 800)
-    }
-
-    // TODO: Detail search Reset
-    function searchbarDetailOptionsReset() {
-        apiJson.ar.min = 0
-        apiJson.ar.max = 0
-        apiJson.cs.min = 0
-        apiJson.cs.max = 0
-        apiJson.accuracy.min = 0
-        apiJson.accuracy.max = 0
-        apiJson.drain.min = 0
-        apiJson.drain.max = 0
-        apiJson.bpm.min = 0
-        apiJson.bpm.max = 0
-        apiJson.difficultyRating.min = 0
-        apiJson.difficultyRating.max = 0
-        detailSearchTmp.ar.min = 0
-        detailSearchTmp.ar.max = 0
-        detailSearchTmp.cs.min = 0
-        detailSearchTmp.cs.max = 0
-        detailSearchTmp.accuracy.min = 0
-        detailSearchTmp.accuracy.max = 0
-        detailSearchTmp.drain.min = 0
-        detailSearchTmp.drain.max = 0
-        detailSearchTmp.bpm.min = 0
-        detailSearchTmp.bpm.max = 0
-        detailSearchTmp.difficultyRating.min = 0
-        detailSearchTmp.difficultyRating.max = 0
-
-        requestNewBeatmapData(false)
-    }
-
     function requestNewBeatmapData(append=true) {
         setGlobalState("musicPlayerIsPlaying", false)
         setGlobalState("musicPlayerIsPaused", true)
@@ -350,183 +82,6 @@ function Searchbar() {
     return (
         <div className="searchbar-block">
             <Search className={"searchbar-input"} onSearch={searchHandler} onChange={searchbarChangeHandler} size="large" enterButton="Search" placeholder="Search...." allowClear="true" value={apiJson.query}/>
-            <ul className="searchbar-type-selector">
-                <li data-active={detailSearch === true ? false : true} onClick={(e) => searchbarTypeHandler(e, false)}>
-                   <p>Search</p> 
-                </li>
-                <li data-active={detailSearch === true ? true : false} onClick={(e) => searchbarTypeHandler(e, true)}>
-                   <p>Detail Search</p> 
-                </li>
-            </ul>
-            <ul className="searchbar-options">
-                <li className="searchbar-option">
-                    <strong>Mode</strong>
-                    <ul>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'm', '')}>
-                            <p data-active={apiJson.m === '' ? true : false}>Any</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'm', '0')}>
-                            <p data-active={apiJson.m === '0' ? true : false}>osu!</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'm', '1')}>
-                            <p data-active={apiJson.m === '1' ? true : false}>osu!taiko</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'm', '2')}>
-                            <p data-active={apiJson.m === '2' ? true : false}>osu!catch</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'm', '3')}>
-                            <p data-active={apiJson.m === '3' ? true : false}>osu!mania</p>
-                        </li>
-                    </ul>
-                </li>
-                <li className="searchbar-option">
-                    <strong>Categories</strong>
-                    <ul>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'ranked', 'any')}>
-                            <p data-active={(apiJson.ranked === 'all') ? true : false}>Any</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'ranked', '')}>
-                            <p data-active={(apiJson.ranked === '' || apiJson.ranked === '') ? true : false}>Has Leaderboard</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'ranked', 'ranked')}>
-                            <p data-active={apiJson.ranked.includes('ranked') ? true : false}>Ranked</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'ranked', 'qualified')}>
-                            <p data-active={apiJson.ranked.includes('qualified') ? true : false}>Qualified</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'ranked', 'loved')}>
-                            <p data-active={apiJson.ranked.includes('loved') ? true : false}>Loved</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'ranked', 'pending')}>
-                            <p data-active={apiJson.ranked.includes('pending') ? true : false}>Pending</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'ranked', 'wip')}>
-                            <p data-active={apiJson.ranked.includes('wip') ? true : false}>Wip</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'ranked', 'graveyard')}>
-                            <p data-active={apiJson.ranked.includes('graveyard') ? true : false}>Graveyard</p>
-                        </li>
-                    </ul>
-                </li>
-                <li className="searchbar-option">
-                    <strong>Explicit Content</strong>
-                    <ul>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'nsfw', false)}>
-                            <p data-active={(apiJson.nsfw === false || apiJson.nsfw === null) ? true : false}>Hide</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandler(e, 'nsfw', true)}>
-                            <p data-active={apiJson.nsfw === true ? true : false}>Show</p>
-                        </li>
-                    </ul>
-                </li>
-                <li className="searchbar-option">
-                    <strong>Extra</strong>
-                    <ul>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForExtra(e, 'video')}>
-                            <p data-active={(apiJson.extra === 'storyboard.video' || apiJson.extra === 'video') ? true : false}>Has Video</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForExtra(e, 'storyboard')}>
-                            <p data-active={apiJson.extra === 'storyboard.video' || apiJson.extra === 'storyboard' ? true : false}>Has Storyboard</p>
-                        </li>
-                    </ul>
-                </li>
-                <li className="searchbar-option">
-                    <strong>Sort by</strong>
-                    <ul>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForSortBy(e, 'title')}>
-                            <p data-active={apiJson.sort.includes('title') ? true : false}>Title</p>
-                            <i data-asc={apiJson.sort.includes('title') && apiJson.sort.includes('asc') ? true : false} className="fa-solid fa-chevron-down" style={{display: (apiJson.sort.includes('title') ? '' : 'none')}}></i>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForSortBy(e, 'artist')}>
-                            <p data-active={apiJson.sort.includes('artist') ? true : false}>Artist</p>
-                            <i data-asc={apiJson.sort.includes('artist') && apiJson.sort.includes('asc') ? true : false} className="fa-solid fa-chevron-down" style={{display: (apiJson.sort.includes('artist') ? '' : 'none')}}></i>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForSortBy(e, 'difficulty')}>
-                            <p data-active={apiJson.sort.includes('difficulty') ? true : false}>Difficulty</p>
-                            <i data-asc={apiJson.sort.includes('difficulty') && apiJson.sort.includes('asc') ? true : false} className="fa-solid fa-chevron-down" style={{display: (apiJson.sort.includes('difficulty') ? '' : 'none')}}></i>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForSortBy(e, 'ranked')}>
-                            <p data-active={apiJson.sort.includes('ranked') || apiJson.sort === '' ? true : false}>Ranked</p>
-                            <i data-asc={apiJson.sort.includes('ranked') && apiJson.sort.includes('asc') ? true : false} className="fa-solid fa-chevron-down" style={{display: (apiJson.sort.includes('ranked') ? '' : 'none')}}></i>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForSortBy(e, 'updated')}>
-                            <p data-active={apiJson.sort.includes('updated') ? true : false}>Updated</p>
-                            <i data-asc={apiJson.sort.includes('updated') && apiJson.sort.includes('asc') ? true : false} className="fa-solid fa-chevron-down" style={{display: (apiJson.sort.includes('updated') ? '' : 'none')}}></i>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForSortBy(e, 'plays')}>
-                            <p data-active={apiJson.sort.includes('plays') ? true : false}>Plays</p>
-                            <i data-asc={apiJson.sort.includes('plays') && apiJson.sort.includes('asc') ? true : false} className="fa-solid fa-chevron-down" style={{display: (apiJson.sort.includes('plays') ? '' : 'none')}}></i>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForSortBy(e, 'favourites')}>
-                            <p data-active={apiJson.sort.includes('favourites') ? true : false}>Favourites</p>
-                            <i data-asc={apiJson.sort.includes('favourites') && apiJson.sort.includes('asc') ? true : false} className="fa-solid fa-chevron-down" style={{display: (apiJson.sort.includes('favourites') ? '' : 'none')}}></i>
-                        </li>
-                    </ul>
-                </li>
-                <li className="searchbar-option">
-                    <strong>Download Options</strong>
-                    <ul>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForDownload(e)}>
-                            <p data-active={globalDirectDownload ? true : false}>Direct Download</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForDlOptions(e, 'video')}>
-                            <p data-active={globalNoVideo ? true : false}>No Video</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForDlOptions(e, 'background')}>
-                            <p data-active={globalNoBg ? true : false}>No Background</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForDlOptions(e, 'hitsound')}>
-                            <p data-active={globalNoHitsound ? true : false}>No Hitsound</p>
-                        </li>
-                        <li onClick={(e) => searchbarOptionChangeHandlerForDlOptions(e, 'storyboard')}>
-                            <p data-active={globalNoStoryboard ? true : false}>No Storyboard</p>
-                        </li>
-                    </ul>
-                </li>
-                <li className="searchbar-option">
-                    <strong>Zip Download (Beta)</strong>
-                    <ul>
-                        <li onClick={(e) => zipDownloadHandler(e)}>
-                            <p>Seleted Download</p>
-                        </li>
-                        <li onClick={(e) => zipDownloadHandler(e, false)}>
-                            <p>All Download</p>
-                        </li>
-                    </ul>
-                </li>
-            </ul>
-            <animated.ul 
-                className="searchbar-detail-options"
-                data-active={detailSearch === true ? true : false}
-                style={{
-                opacity: x.to({ range: [0, 1], output: [0, 1] }),
-                }}
-            >
-                <li className="searchbar-option">
-                    <strong>AR</strong>
-                    <Slider range step={0.1} onAfterChange={detailSearchbarValueAfterChangeHandler} onChange={detailSearchbarValueHandlerForAR} defaultValue={[0, 10]} max={10}/>
-                </li>
-                <li className="searchbar-option">
-                    <strong>CS</strong>
-                    <Slider range step={0.1} onAfterChange={detailSearchbarValueAfterChangeHandler} onChange={detailSearchbarValueHandlerForCS} defaultValue={[0, 10]} max={10}/>
-                </li>
-                <li className="searchbar-option">
-                    <strong>OD</strong>
-                    <Slider range step={0.1} onAfterChange={detailSearchbarValueAfterChangeHandler} onChange={detailSearchbarValueHandlerForOD} defaultValue={[0, 10]} max={10}/>
-                </li>
-                <li className="searchbar-option">
-                    <strong>HP</strong>
-                    <Slider range step={0.1} onAfterChange={detailSearchbarValueAfterChangeHandler} onChange={detailSearchbarValueHandlerForHP} defaultValue={[0, 10]} max={10}/>
-                </li>
-                <li className="searchbar-option">
-                    <strong>BPM</strong>
-                    <Slider range step={1} onAfterChange={detailSearchbarValueAfterChangeHandler} onChange={detailSearchbarValueHandlerForBPM} defaultValue={[0, 500]} max={500}/>
-                </li>
-                <li className="searchbar-option">
-                    <strong>Star Rating</strong>
-                    <Slider range step={0.1} onAfterChange={detailSearchbarValueAfterChangeHandler} onChange={detailSearchbarValueHandlerForSR} defaultValue={[0, 11]} max={11}/>
-                </li>
-            </animated.ul>
             {/* <p className="searchbar-detail-reset" data-show={detailSearchChange === true ? true : false} onClick={(e) => searchbarDetailOptionsReset()}>Reset</p> */}
         </div>
     )
