+++ src/Pages/index.js
@@ -1,6 +1,7 @@
 import Beatmaps from './Beatmaps'
 import Download from './Download'
 import Info from './Info'
+import NotFound from './404'
 import TsetD from './TsetD'
 
-export { Beatmaps, Download, Info, TsetD }
+export { Beatmaps, Download, Info, NotFound, TsetD }
