+++ src/assets/css/global.css
@@ -231,6 +231,7 @@ div.container { margin: 0 auto; max-width: 70vw; padding: 0 0 var(--footer-heigh
     footer { font-size: 10px !important; flex-direction: column-reverse !important; align-items: center !important; }
     footer > ul { margin-left: 0 !important; margin-bottom: 0 !important; }
     div.container { max-width: 85vw !important; }
+    div.settings-area { width: 60vw !important; max-width: 60vw; transform: translateX(-60vw) translateY(-50%); }
     ul.searchbar-type-selector li { width: 300px !important; }
     ul.beatmap-list > li { width: 100% !important; }
     .portal-info-left { margin-bottom: 1em !important; }
@@ -239,12 +240,17 @@ div.container { margin: 0 auto; max-width: 70vw; padding: 0 0 var(--footer-heigh
     div.more-info-two-side > ul.desc { flex-direction: column; }
 }
 @media (max-width: 860px) {
+    div.settings-area { width: 80vw !important; max-width: 80vw; transform: translateX(-80vw) translateY(-50%); }
     ul.searchbar-type-selector li { width: 200px !important; }
     div.info-Single { width: 80vw !important; }
     div.more-info-two-side { flex-direction: column; align-items: center; }
     ul.portal-buttons { flex-direction: row;}
 }
 @media (max-width: 501px) {
+    div.settings-area { width: 100vw !important; max-width: 100vw; transform: translateX(-100vw) translateY(-50%); max-height: 80vh; overflow-y: hidden; }
+    div.settings-area ul.settings li.settings-option strong { font-size: 12px !important; }
+    div.settings-area ul.settings li.settings-option ul { font-size: 10px !important; }
+    div.settings-area ul.settings li.settings-option:first-child { display: none !important; }
     /* nav div.expand { visibility: hidden !important; display: none !important; } */
     /* nav div.collapsed { visibility: visible !important; display: block !important; } */
     ul.searchbar-type-selector li { width: 140px !important; }
