+++ src/store/index.js
@@ -3,6 +3,7 @@ import { createGlobalState } from "react-hooks-global-state"
 const { setGlobalState, useGlobalState, getGlobalState } = createGlobalState({
     firstLoad: true,
     loading: true,
+    settingTab: false,
     apiErrorCount: 0,
     noResult: false,
     notification: {},
